use id(Name, Setting) structures for lp names.
It turns out that lifepaths dont need to have unique names across different settings. For example Groom appears three times: Villager, City, and Noble Court. These lifepaths differ in detail as well.

As a result we need to uniquely identify lifepaths across settings: hence the id structure.

This also includes a small quality of life improvement that allows the user to continue to call character_path with bare names and omit the setting if they dont need to disambiguate them.